requirements.txt
asgiref==3.7.2
Django==3.2.16
djangorestframework==3.12.4
pytz==2023.3.post1
sqlparse==0.4.4
